\name{ecespa}
\alias{ecespa}
\alias{ecespa}
\docType{package}
\title{
Functions for spatial point pattern analysis

}
\description{
Some wrappers, functions and data for spatial point pattern analysis. 
}
\details{
\tabular{ll}{
Package: \tab ecespa\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2007-10-22\cr
License: \tab GPL (version 2 or later)\cr
}
}
\author{
Marcelino de la Cruz Rot, with contributions of Philip M. Dixon and heavily borrowing  Baddeley's & Turner's spatstat code .
}


\keyword{ package }
